split SetSearchSpaceBasedTuningStrategy into .cpp and .h and remove inline
